# docs/openapi.json
{
  "components": {
    "schemas": {
      "AgentCloneRequest": {
        "description": "`POST /agents/{id}/clone`（3.2.2）。",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "AgentCloneRequest",
        "type": "object"
      },
      "AgentCreate": {
        "description": "3.3.1 的请求体（`tool_ids` / `knowledge_base_ids` / `workflow_id` 在 Phase 1 必须为空）。",
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "is_template": {
            "default": false,
            "title": "Is Template",
            "type": "boolean"
          },
          "knowledge_base_ids": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Base Ids",
            "type": "array"
          },
          "max_steps": {
            "default": 8,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "memory_config": {
            "$ref": "#/components/schemas/MemoryConfigDTO"
          },
          "model_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "model_params": {
            "additionalProperties": true,
            "title": "Model Params",
            "type": "object"
          },
          "model_provider_id": {
            "maxLength": 26,
            "minLength": 26,
            "title": "Model Provider Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "status": {
            "default": "enabled",
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Status",
            "type": "string"
          },
          "system_prompt": {
            "default": "",
            "title": "System Prompt",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "timeout_seconds": {
            "default": 180,
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "type": "integer"
          },
          "tool_ids": {
            "items": {
              "type": "string"
            },
            "title": "Tool Ids",
            "type": "array"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          }
        },
        "required": [
          "name",
          "model_provider_id",
          "model_name"
        ],
        "title": "AgentCreate",
        "type": "object"
      },
...
      "AgentUpdate": {
        "description": "`PATCH`：只传需要改的字段；`system_prompt` 变化会写 `agent_prompt_versions`（2.4）。",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "max_steps": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          },
          "memory_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryConfigDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Params"
          },
          "model_provider_id": {
            "anyOf": [
              {
                "maxLength": 26,
                "minLength": 26,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Provider Id"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "prompt_note": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Note"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "enabled",
                  "disabled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "maximum": 3600.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "title": "AgentUpdate",
        "type": "object"
      },
      "ApiResponse_AgentRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentRead"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[AgentRead]",
        "type": "object"
      },
...
      "ApiResponse_ProviderRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProviderRead"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ProviderRead]",
        "type": "object"
      },
      "ApiResponse_ProviderTestResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProviderTestResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ProviderTestResult]",
        "type": "object"
      },
...
      "ApiResponse_list_AgentPromptVersionRead__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AgentPromptVersionRead"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[AgentPromptVersionRead]]",
        "type": "object"
      },
      "ApiResponse_list_AgentRead__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AgentRead"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[AgentRead]]",
        "type": "object"
      },
...
      "ApiResponse_list_ProviderRead__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProviderRead"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[ProviderRead]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryConfigDTO": {
        "properties": {
          "long_term": {
            "$ref": "#/components/schemas/LongTermConfigDTO"
          },
          "short_term": {
            "$ref": "#/components/schemas/ShortTermConfigDTO"
          }
        },
        "title": "MemoryConfigDTO",
        "type": "object"
      },
...
      "ProviderCreate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Base Url",
            "type": "string"
          },
          "default_model": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model"
          },
          "default_params": {
            "additionalProperties": true,
            "title": "Default Params",
            "type": "object"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind",
            "default": "openai_compatible"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelEntry"
            },
            "title": "Models",
            "type": "array"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "status": {
            "default": "enabled",
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "base_url"
        ],
        "title": "ProviderCreate",
        "type": "object"
      },
      "ProviderKind": {
        "description": "LLM Provider 类型（4.1.2）。",
        "enum": [
          "openai_compatible"
        ],
        "title": "ProviderKind",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "配置驱动的 LLM Agent 编排与执行平台（详细设计 3.x）",
    "title": "Agent Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agents": {
      "get": {
        "operationId": "list_agents_api_v1_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AgentRead__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent 列表（?q=&status=&tag=）",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_agent_api_v1_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AgentRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "新建 Agent（校验引用与白名单）",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}": {
      "delete": {
        "operationId": "delete_agent_api_v1_agents__agent_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "软删除 Agent",
        "tags": [
          "agents"
        ]
      },
      "get": {
        "operationId": "get_agent_api_v1_agents__agent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AgentRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent 详情",
        "tags": [
          "agents"
        ]
      },
      "patch": {
        "operationId": "update_agent_api_v1_agents__agent_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AgentRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "更新 Agent（prompt 变更留档）",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}/clone": {
      "post": {
        "operationId": "clone_agent_api_v1_agents__agent_id__clone_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCloneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AgentRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "复制为新 Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_id}/prompt-versions": {
      "get": {
        "operationId": "list_prompt_versions_api_v1_agents__agent_id__prompt_versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AgentPromptVersionRead__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prompt 历史版本",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/v1/model-providers": {
      "get": {
        "operationId": "list_model_providers_api_v1_model_providers_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ProviderRead__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider 列表",
        "tags": [
          "models"
        ]
      },
      "post": {
        "operationId": "create_model_provider_api_v1_model_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProviderRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "新建 Provider（api_key 加密存储）",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/model-providers/{provider_id}": {
      "delete": {
        "operationId": "delete_model_provider_api_v1_model_providers__provider_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "删除 Provider（被引用 → 409）",
        "tags": [
          "models"
        ]
      },
      "get": {
        "operationId": "get_model_provider_api_v1_model_providers__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProviderRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider 详情",
        "tags": [
          "models"
        ]
      },
      "patch": {
        "operationId": "update_model_provider_api_v1_model_providers__provider_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProviderRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "更新 Provider",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/model-providers/{provider_id}/test": {
      "post": {
        "operationId": "test_model_provider_api_v1_model_providers__provider_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProviderTestResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "连通性检测（1-token）",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}